1.添加推送功能。 2.修复钱包bug。 3.添加帮他人预约功能。

+++ app/src/main/java/com/zjc/drivingschool/db/model/AccountBalance.java
@@ -10,13 +10,13 @@ import java.math.BigDecimal;
  * @date 2016年7月21日
  */
 public class AccountBalance implements Serializable {
-    private int balance;// 余额
+    private double balance;// 余额
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
